What I did Apply port speed again after disabling port AN.
In PortsOrch, when autonegotiation is disabled for a port, this PR ensures
that the port speed is updated correctly. If a new speed is configured while
autonegotiation is disabled, the new speed is applied; otherwise, the current
speed setting is reconfigured.

Why I did it
In PortsOrch, the port speed configuration will not be set to SAI when port AN is enabled.
Changing the port speed while port AN is enabled will not take effect after port AN is disabled.

In some chips, the port speed can not be modified when port AN is enabled.

How I verified it

Enable port AN.
Set port speed from speed1 to speed2
Disable port AN.
Make sure port speed is speed2 instead of speed1.